implement all keypresses (numbers and operations)
this is ugly as shit but it does work. Included a refactor to make it so
maps are from textcontent -> action rather than button to action.

a future refactor should separate the calculator and the DOM and the
keypressing into separate controllers because this is a bit of a
hilarious mess at the moment.